New ActorRel interface. Moved GrassActor to the new style
