+++ src/j3d/WindWaker/Actors.ts
@@ -18,8 +18,6 @@ import ArrayBufferSlice from '../../ArrayBufferSlice';
 import { colorFromRGBA } from '../../Color';
 import { GfxRenderInstManager, GfxRendererLayer } from '../../gfx/render/GfxRenderer';
 
-import { AGrass } from './Grass';
-
 export interface ActorInfo { 
     relName: string, 
     subtype: number, 
@@ -37,6 +35,12 @@ export interface Actor {
     rotationY: number;
 };
 
+// The REL table maps .rel names to our implementations
+interface ActorRel {
+    // @TODO: Most actors have draw and update functions
+}
+const kRelTable: { [relName: string]: { new (context: WindWakerRenderer, actor: Actor): ActorRel } } = {};
+
 // Special-case actors
 
 export const enum LightTevColorType {
@@ -47,8 +51,10 @@ export const enum LightTevColorType {
     BG3 = 4,
 }
 
-const scratchMatrix = mat4.create();
+const scratchMat4a = mat4.create();
+const scratchMat4b = mat4.create();
 const scratchVec3a = vec3.create();
+const scratchVec3b = vec3.create();
 
 // dScnKy_env_light_c::settingTevStruct
 export function settingTevStruct(actor: J3DModelInstanceSimple, type: LightTevColorType, colors: KyankoColors): void {
@@ -1313,6 +1319,12 @@ export async function loadActor(device: GfxDevice, renderer: WindWakerRenderer, 
         // TODO(jstpierre): Scale, Rotation
         return emitter;
     }
+
+    const relConstructor = kRelTable[actor.info.relName];
+    if (relConstructor) {
+        const actorObj = new relConstructor(renderer, actor);
+        return;
+    }
     
     // Generic Torch
     if (actor.name === 'bonbori') {
@@ -1323,13 +1335,13 @@ export async function loadActor(device: GfxDevice, renderer: WindWakerRenderer, 
         if (type === 0x3F)
             type = 0;
 
-        setModelMatrix(scratchMatrix);
+        setModelMatrix(scratchMat4a);
         vec3.set(scratchVec3a, 0, 0, 0);
         if (type === 0 || type === 3) {
             const m = buildModel(rarc, obm ? `bdl/obm_shokudai1.bdl` : `bdl/vktsd.bdl`);
             scratchVec3a[1] += 140;
         }
-        vec3.transformMat4(scratchVec3a, scratchVec3a, scratchMatrix);
+        vec3.transformMat4(scratchVec3a, scratchVec3a, scratchMat4a);
 
         // Create particle systems.
         const pa = createEmitter(0x0001);
@@ -1377,17 +1389,6 @@ export async function loadActor(device: GfxDevice, renderer: WindWakerRenderer, 
     }
     // Under-water treasure points. Perhaps spawn at some point?
     // else if (actor.name === 'Salvage' || actor.name === 'Salvag2' || actor.name === 'SalvagE' || actor.name === 'SalvagN' || actor.name === 'SalvFM') return;
-    // Grass/Flowers/Small Trees. Procedurally generated by the engine.
-    // https://github.com/LagoLunatic/WW-Hacking-Docs/blob/6e1ecdadbdf5124e7f6ff037106deb29a5f7238b/Entity%20DZx%20Formats.txt#L695
-    else if (
-        actor.name === 'kusax1' || actor.name === 'kusax7'  || actor.name === 'kusax21' ||
-        actor.name === 'flower' || actor.name === 'flwr7'   || actor.name === 'flwr17' ||
-        actor.name === 'pflower'|| actor.name === 'pflwrx7' || 
-        actor.name === 'swood'  || actor.name === 'swood3'  || actor.name === 'swood5'
-    ) {
-        if (actor.layer === -1 || actor.layer === 0)
-            return AGrass.create(renderer, actor);
-    }
     // // Bushes. Procedurally generated by the engine.
     // else if (actor.name === 'woodb' || actor.name === 'woodbx') return;
     // // Rope. Procedurally generated by the engine.
@@ -1430,3 +1431,158 @@ export async function loadActor(device: GfxDevice, renderer: WindWakerRenderer, 
         else console.warn(`Unknown object: ${actor.name} / ${roomRenderer.name} Layer ${actor.layer} / ${hexzero(actor.parameters, 8)}`);
     }
 }
+
+// ---------------------------------------------
+// Grass Actor
+// ---------------------------------------------
+class AGrass implements ActorRel {
+    static kSpawnPatterns = [
+        { group: 0, count: 1 },
+        { group: 0, count: 7 },
+        { group: 1, count: 15 },
+        { group: 2, count: 3 },
+        { group: 3, count: 7 },
+        { group: 4, count: 11 },
+        { group: 5, count: 7 },
+        { group: 6, count: 5 },
+    ];
+    
+    static kSpawnOffsets = [
+        [
+            [0, 0, 0],
+            [3, 0, -50],
+            [-2, 0, 50],
+            [50, 0, 27],
+            [52, 0, -25],
+            [-50, 0, 22],
+            [-50, 0, -29],
+        ],
+        [
+            [-18, 0, 76],
+            [-15, 0, 26],
+            [133, 0, 0],
+            [80, 0, 23],
+            [86, 0, -83],
+            [33, 0, -56],
+            [83, 0, -27],
+            [-120, 0, -26],
+            [-18, 0, -65],
+            [-20, 0, -21],
+            [-73, 0, 1],
+            [-67, 0, -102],
+            [-21, 0, 126],
+            [-120, 0, -78],
+            [-70, 0, -49],
+            [32, 0, 103],
+            [34, 0, 51],
+            [-72, 0, 98],
+            [-68, 0, 47],
+            [33, 0, -5],
+            [135, 0, -53],
+        ],
+        [
+            [-75, 0, -50],
+            [75, 0, -25],
+            [14, 0, 106],
+        ],
+        [
+            [-24, 0, -28],
+            [27, 0, -28],
+            [-21, 0, 33],
+            [-18, 0, -34],
+            [44, 0, -4],
+            [41, 0, 10],
+            [24, 0, 39],
+        ],
+        [
+            [-55, 0, -22],
+            [-28, 0, -50],
+            [-77, 0, 11],
+            [55, 0, -44],
+            [83, 0, -71],
+            [11, 0, -48],
+            [97, 0, -34],
+            [-74, 0, -57],
+            [31, 0, 58],
+            [59, 0, 30],
+            [13, 0, 23],
+            [-12, 0, 54],
+            [55, 0, 97],
+            [10, 0, 92],
+            [33, 0, -10],
+            [-99, 0, -27],
+            [40, 0, -87],
+        ],
+        [
+            [0, 0, 3],
+            [-26, 0, -29],
+            [7, 0, -25],
+            [31, 0, -5],
+            [-7, 0, 40],
+            [-35, 0, 15],
+            [23, 0, 32],
+        ],
+        [
+            [-40, 0, 0],
+            [0, 0, 0],
+            [80, 0, 0],
+            [-80, 0, 0],
+            [40, 0, 0],
+        ]
+    ];
+
+    constructor(context: WindWakerRenderer, actor: Actor) {
+        const enum FoliageType {
+            Grass,
+            Tree,
+            WhiteFlower,
+            PinkFlower
+        };
+
+        const spawnPatternId = (actor.parameters & 0x00F) >> 0;
+        const type: FoliageType = (actor.parameters & 0x030) >> 4;
+        const itemIdx = (actor.parameters >> 6) & 0x3f; // Determines which item spawns when this is cut down
+
+        const pattern = AGrass.kSpawnPatterns[spawnPatternId];
+        const offsets = AGrass.kSpawnOffsets[pattern.group];
+        const count = pattern.count;
+
+        switch (type) {
+            case FoliageType.Grass:
+                for (let j = 0; j < count; j++) {
+                    // @NOTE: Grass does not observe actor rotation or scale
+                    const offset = vec3.set(scratchVec3a, offsets[j][0], offsets[j][1], offsets[j][2]);
+                    const pos = vec3.add(scratchVec3a, offset, actor.pos);
+                    context.grassPacket.newData(pos, actor.roomIndex, itemIdx);
+                }
+            break;
+
+            case FoliageType.Tree:
+                const rotation = mat4.fromYRotation(scratchMat4a, actor.rotationY);
+
+                for (let j = 0; j < count; j++) {
+                    const offset = vec3.transformMat4(scratchVec3a, offsets[j], rotation);
+                    const pos = vec3.add(scratchVec3b, offset, actor.pos);
+                    context.treePacket.newData(pos, 0, actor.roomIndex);
+                }
+            break;
+
+            case FoliageType.WhiteFlower:
+            case FoliageType.PinkFlower:
+                for (let j = 0; j < count; j++) {
+                    const isPink = (type === FoliageType.PinkFlower);
+
+                    // @NOTE: Flowers do not observe actor rotation or scale
+                    const offset = vec3.set(scratchVec3a, offsets[j][0], offsets[j][1], offsets[j][2]);
+                    const pos = vec3.add(scratchVec3a, offset, actor.pos);
+                    context.flowerPacket.newData(pos, isPink, actor.roomIndex, itemIdx);
+                }
+            break;
+            default:
+                console.warn('Unknown grass actor type');
+        }
+
+        return this;
+    }
+}
+kRelTable['d_a_grass'] = AGrass;
