+++ src/j3d/WindWaker/Grass.ts
@@ -313,11 +313,12 @@ export class FlowerPacket {
         }
     }
 
-    public newData(pos: vec3, type: FlowerType, roomIdx: number, itemIdx: number): FlowerData {
+    public newData(pos: vec3, isPink: boolean, roomIdx: number, itemIdx: number): FlowerData {
         const animIdx = Math.floor(Math.random() * 8);
+        let type = isPink ? FlowerType.PINK : FlowerType.WHITE;
 
         // Island 0x21 uses the Bessou flower (the game does this check here as well)
-        if (this.context.stage === 'sea' && roomIdx === 0x21 && type === FlowerType.PINK) {
+        if (this.context.stage === 'sea' && roomIdx === 0x21 && isPink) {
             type = FlowerType.BESSOU;
         }
 
@@ -1051,157 +1052,5 @@ export class GrassPacket {
             }
         }
         renderInstManager.popTemplateRenderInst();
-    }
-}
-
-// ---------------------------------------------
-// Grass Actor
-// ---------------------------------------------
-const kGrassSpawnPatterns = [
-    { group: 0, count: 1 },
-    { group: 0, count: 7 },
-    { group: 1, count: 15 },
-    { group: 2, count: 3 },
-    { group: 3, count: 7 },
-    { group: 4, count: 11 },
-    { group: 5, count: 7 },
-    { group: 6, count: 5 },
-];
-
-const kGrassSpawnOffsets = [
-    [
-        [0, 0, 0],
-        [3, 0, -50],
-        [-2, 0, 50],
-        [50, 0, 27],
-        [52, 0, -25],
-        [-50, 0, 22],
-        [-50, 0, -29],
-    ],
-    [
-        [-18, 0, 76],
-        [-15, 0, 26],
-        [133, 0, 0],
-        [80, 0, 23],
-        [86, 0, -83],
-        [33, 0, -56],
-        [83, 0, -27],
-        [-120, 0, -26],
-        [-18, 0, -65],
-        [-20, 0, -21],
-        [-73, 0, 1],
-        [-67, 0, -102],
-        [-21, 0, 126],
-        [-120, 0, -78],
-        [-70, 0, -49],
-        [32, 0, 103],
-        [34, 0, 51],
-        [-72, 0, 98],
-        [-68, 0, 47],
-        [33, 0, -5],
-        [135, 0, -53],
-    ],
-    [
-        [-75, 0, -50],
-        [75, 0, -25],
-        [14, 0, 106],
-    ],
-    [
-        [-24, 0, -28],
-        [27, 0, -28],
-        [-21, 0, 33],
-        [-18, 0, -34],
-        [44, 0, -4],
-        [41, 0, 10],
-        [24, 0, 39],
-    ],
-    [
-        [-55, 0, -22],
-        [-28, 0, -50],
-        [-77, 0, 11],
-        [55, 0, -44],
-        [83, 0, -71],
-        [11, 0, -48],
-        [97, 0, -34],
-        [-74, 0, -57],
-        [31, 0, 58],
-        [59, 0, 30],
-        [13, 0, 23],
-        [-12, 0, 54],
-        [55, 0, 97],
-        [10, 0, 92],
-        [33, 0, -10],
-        [-99, 0, -27],
-        [40, 0, -87],
-    ],
-    [
-        [0, 0, 3],
-        [-26, 0, -29],
-        [7, 0, -25],
-        [31, 0, -5],
-        [-7, 0, 40],
-        [-35, 0, 15],
-        [23, 0, 32],
-    ],
-    [
-        [-40, 0, 0],
-        [0, 0, 0],
-        [80, 0, 0],
-        [-80, 0, 0],
-        [40, 0, 0],
-    ]
-];
-
-export class AGrass {
-    public static create(context: WindWakerRenderer, actor: Actor): void {
-        const enum FoliageType {
-            Grass,
-            Tree,
-            WhiteFlower,
-            PinkFlower
-        };
-
-        const spawnPatternId = (actor.parameters & 0x00F) >> 0;
-        const type: FoliageType = (actor.parameters & 0x030) >> 4;
-        const itemIdx = (actor.parameters >> 6) & 0x3f; // Determines which item spawns when this is cut down
-
-        const pattern = kGrassSpawnPatterns[spawnPatternId];
-        const offsets = kGrassSpawnOffsets[pattern.group];
-        const count = pattern.count;
-
-        switch (type) {
-            case FoliageType.Grass:
-                for (let j = 0; j < count; j++) {
-                    // @NOTE: Grass does not observe actor rotation or scale
-                    const offset = vec3.set(scratchVec3a, offsets[j][0], offsets[j][1], offsets[j][2]);
-                    const pos = vec3.add(scratchVec3a, offset, actor.pos);
-                    context.grassPacket.newData(pos, actor.roomIndex, itemIdx);
-                }
-            break;
-
-            case FoliageType.Tree:
-                const rotation = mat4.fromYRotation(scratchMat4a, actor.rotationY);
-
-                for (let j = 0; j < count; j++) {
-                    const offset = vec3.transformMat4(scratchVec3a, offsets[j], rotation);
-                    const pos = vec3.add(scratchVec3b, offset, actor.pos);
-                    context.treePacket.newData(pos, 0, actor.roomIndex);
-                }
-            break;
-
-            case FoliageType.WhiteFlower:
-            case FoliageType.PinkFlower:
-                for (let j = 0; j < count; j++) {
-                    const flowerType = (type === FoliageType.WhiteFlower) ? FlowerType.WHITE : FlowerType.PINK;
-
-                    // @NOTE: Flowers do not observe actor rotation or scale
-                    const offset = vec3.set(scratchVec3a, offsets[j][0], offsets[j][1], offsets[j][2]);
-                    const pos = vec3.add(scratchVec3a, offset, actor.pos);
-                    context.flowerPacket.newData(pos, flowerType, actor.roomIndex, itemIdx);
-                }
-            break;
-            default:
-                console.warn('Unknown grass actor type');
-        }
     }
 }
